requirements.txt
Django>=5.2,<5.3
django-environ>=0.11,<1.0
